geom2d: introducing add and sub operators
