+++ foxes/utils/geom2d/__init__.py
@@ -1,5 +1,4 @@
 from .area_geometry import AreaGeometry, InvertedAreaGeometry
+from .area_geometry import AreaUnion, AreaIntersection
 from .polygon import ClosedPolygon
 from .circle import Circle
-from .union import AreaUnion
-from .intersection import AreaIntersection
